Added kernel re-entry, and the locking required for the serial interrupt
Hardware interrupts are now enabled when handling syscalls, which allows
the serial interrupt to be handled during a syscall.  Rescheduling of
processes is skipped when a syscall is in progress, but in future it will
only be delayed until the end of the current syscall.  The only things
that need to be locked, as a result of this, are things that are also
manipulated by the serial interrupt, namely the serial buffer operations
and the resuming of suspended processes due to available serial data.