Fixed elf loading bug and added relocation
The previous elf loader was only loading one segment instead of all program
header load segments, which was causing the process segment to be way too
small, pushing the bss out into neighbouring memory.  I believe this was
causing the issue where the command list data was overwriting the text
segment causing a catastrophic failure.

I also fixed an issue where a process could fail after modifying the memory
maps, resulting in an attempt to run uninitialized memory as code, which
caused an illegal instruction execption, but could have been worse.  The fix
is a hack, but if the EKILL error is returned, the syscall handler will kill
the process instead of returning the error.

I also added relocation when loading segments.  If there is a GNU_RELRO
segment, it will intepret the segment as a list of pointers, and add the
text segment starting location to the existing contents.  It works for the
pre-initialized string data in the date command.  This solution wont work
with the .got section though, at least not yet