Fixed tty bug where characters were lost
The input processor was reading in as many characters as possible,
which included characters after the linefeed, but the read() function
wasn't checking for the linefeed, and including the extra characters
in the data read during ICANON mode.  They would then be lost.

A change to init is also included, which tries to execute a copy of
sh on disk, and then optionally falls back to the builtin version