Added initial networking module
This is the initial commit of the networking subsystem that I've been
working on for the past two weeks.  It supports very simple UDP messages
without checksum calculations or address processing, and is able to
receive UDP messages sent from the host computer.  Sending shows the
packets in wireshark, but receiving on the host is not quite working.
I wanted to get the partial work committed before making changes to
implement a better version of connection handling.