Fixed process not resuming after write buffer full
This was the actual cause of the directory listing hanging on large
directories after increasing the clock speed to 25MHz.  I had
incorrectly chased a logic bug in the CPLD, but it was actually this
long standing bug that was identified in comments in pipe.c.  The issue
with pipes still exists, but at least tty resume is handled for now.

A nicer solution for the future would either be a hook that the
hardware-specific tty driver can call on the tty parent driver, or a
nicer way of suspending and resuming processes instead of the
syscall-specific rdev number method