Fixed the issue that was making it break
It turned out to be the removed SAVE_SYSCALL/RESTORE_SYSCALL in the
enter_syscall function that caused it to not work, at least where it was
switching alright according to the LED but not getting to the prompt.
Saving the syscall only in SAVE_CONTEXT wasn't working because the stack
was swapped before the stack pointer was saved into current_syscall.
Since it's not even a stack argument, it could be set right after the
SAVE_SYSCALL in SAVE_CONTEXT