Fixed the issue where it was crashing on soon after boot
It was caused by the reordered members of struct process, but making the
bitflags be a long word (and or moving the mem map around a bit) seems
to have fixed the problem.  I assume it was some kind of alignment or
padding issue

Once that was fixed, it would start but hang while task switching, which
was the return value being incorrect.  It even called waitpid (#7)
twice.  I was able to implement the better return value setting code,
which saves %d0 on entry from user space only, with the C code only ever
setting the value in the process struct instead of in the context
directly.  At first, it wasn't working because I had the inital d0 save
in enter_irq, but that would overwrite it when re-entering through an
interrupt