Fixed a bug where it crashed to rom when exiting
It only affected socket server processes, and it wouldn't have a stack
trace/fatal error.  It would just end up right in the rom.  It turned
out to be a double free on an invalid file pointer due to the suspend
in the accept() syscall not NULLing the pointer while also returning 0