Finally removed the hardcoded serial irq
There's now two tables, the real vector table with all entries mapped
to the generic kernel entry functions, and a second table of C function
pointers that handle a driver-registered irq.  The `request_irq`
function adds callbacks to the table, and while it doesn't support it at
the moment, it would be possible to have irq sharing using chaining,
like the linux kernel has.