Added separate user and kernel stack when using user mode
This commit includes a bunch of changes that have accumulated over the
last few months.  The biggest ones are that a separate stack was added
when user mode is enabled, for a per-process user stack.  I also fixed a
bug where the MMU enable flag was wrong, so it never actually enabled
the MMU.  Now that it's enabled, it causes many errors and won't boot,
but that's because the memory segments from the original process that is
forked are not mapped to the new process's address space, so they cause
a protection fault when accessed.  It's still using the non-MMU segment
copying code, which must be modified, but that will require refactoring
memory segments a bit to allow copy-on-write and loading on page fault.