Refactored process memory for MMU
This is a bunch of changes that I've had around for months now.  I
removed memory_object and replaced it with either the vfile when MMU is
enabled, or a memory_region object when it's disabled.  The memory
region is refcounted like before, and gets freed in one block, but the
MMU-enabled mode will now rely on the per-page refcount.  It should work
with MMU disabled, but still doesn't work when it's enabled