Fixed bug where process would run as pid 0
If a process was cloned or created by init, and failed, it would call
close_proc() instead of exit_proc() which would leave it on the queue
despite being freed.  It would then continue to run, and cause a null
pointer exception in MMU mode, but probably would just cycle in non-MMU
mode.