Found and fixed major bug with exec()
When a process was calling the exec() syscall, the process was not being
properly reset, so open files would still be open, registered signals
would still be active, and possibly wait events might still be active.
The thing that actually uncovered the issue was that the extended
context switch wasn't happening after exec(), so the previous process's
mmu table was still active instead of the newly created one, which
sometimes caused a spurious page fault when no page faults should have
been occurring.

I've added a reset_proc() function that is called by load_binary() in
binaries.c, which does all the resets except for the extended switch,
wihch is manually called in the exec.c functions, but should eventually
be moved to a better place.

In the process, I also found a bug where fd_tables would not be properly
freed and deallocated based on the refcount that was previously added.

A few other small bug fixes are included as well