Added mmu address translations using usercopy and iovec
This is a major refactoring of the API to add the new iovec and
iovec_iter types to the read/write functions, so that the user buffers
don't have to be copied to an intermediate buffer, but can invisibly be
used by all file and driver interfaces.

For other values passed from user space, it's using intermediate buffers
because it's easier for now.  I'd like to improve it later, but it
should work for now.

This also adds the m68k functions to copy data to and from user space
using the MOVES instruction, which I *think* should use the CRP register
and user ATC/TLB entries in the MMU instead of the kernel space.  I
haven't fully tested it though, and enabling features still seems to
break, so I'll debug that next.

There are probably a lot of bugs introduced by this commit, but it's so
big, I can't really catch them all now.  I'll have to test this
thoroughly over the next while and try to fix what I can